BREAKING: Added support for user preferences
Now plugins can define preferences per-user as a schema, and the user can modify them at will. This will allow customizing the UI and backend server operations for each user.

This is a large update that includes a bunch of changes/refactors/fixes that were implemented while setting up the preferences.