App Settings Schema Validation
The app settings schema was not validated, which meant that the user could technically put any value into the settings field of an app. This fix makes sure that the settings and settings schema are always compatible. Fixes #347.

Also fixed a security issue with object meta updates - adversarial users/apps could mess with the database because updates were not correctly escaped. No more.