ServiceWorker and Development Mode
Heedy relies on having compatible and latest versions of all plugin modules used by the frontend. In order to be fast, it also requires extensive caching. This led to issues getting browsers to see the lastest code edits, or to correctly load updated plugin code. With the introduction of the serviceworker, this problem is fixed. Browsers will now cache everything per run of heedy, and refreshing their cache each time heedy is restarted, to make sure plugins and all code is at the latest version.

This is not good for debugging code, where you might want to edit code during runtime. For that reason, a development mode was added, with the --development flag, which sets the serviceworker to always use the network, and not cache anything, effectively disabling all caching.