Heedy frontend is now ServiceWorker aware
The frontend now asks to be refreshed if the serviceworker is updated. It also updates the worker whenever the websocket is reconnected.
Restarting the server to apply updates now also waits for the ServiceWorker to finish installing before navigating.

The final challenge is non-https use over local network, where all the existing problems are still there, because sw doesn't work in non-secure contexts.