BUGFIX: Cache-Control headers
Static files had no caching policy, which was leading to a ton of issues with updates, even with serviceworker!!

On the other hand, rest api had no-cache, when it really needed no-store.

This sets up the cache for static files as no-cache in dev mode, and with stale-while-revalidate for a week for normal mode, with a 10s fresh cache.
This seems like a good compromise of fast load time, but keeping the content up-to-date on refreshes. This also fixes the serviceworker issue of using previously cached resources.