BUGFIX: no-cache static wherever the serviceworker is used
Over http the serviceworker isnt used, except on localhost, so we really do want cache headers.
However, if we have cached stuff, the serviceworker will use the cached stuff instead of the new stuff!!!

So over https and localhost, no-cache headers are used, and the serviceworker handles caching, otherwise, stale-while-revalidate caching is used for cases when there is no serviceworker.