Static server now checks for standard files before precompressed
There were problems when building production versions of plugins, which created precompressed frontend assets. When a new debug build was added, the files would not be seen, since the frontend would still use the precompressed assets. This was now changed, with the opposite functionality - if original files are present, they are used in first priority. If not, only then does it look for pre-compressed assets. This allows useful debugging, but in production builds, original files are removed, so the server uses pre-compressed.