WIP: Custom Visualizations, UI
Work in progress for enabling custom visualizations.
The UI for custom visualizations (creating and editing) is functional. The actual visualization code that will run these custom visualizations is not implemented.

Since this is a large refactor of the entire visualization subsystem, all visualizations other than table are currently disabled as the corresponding API is built.