WIP: Custom visualizations - linechart
The linechart was the first real test of the templating system. After some work, I have pretty much identical functionality in the new line chart as the old one, but with this one being based entirely on the configuration templates rather than a custom preprocessor.

This commit also upgrades chartjs to the latest version, and adds date-fns, which will be the replacement to momentjs for heedy.